+++ routes/web.php
@@ -1,9 +1,10 @@
 <?php
 
-use App\Http\Controllers\ListingController;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -19,7 +20,7 @@ use Illuminate\Support\Facades\Route;
 //Naming Conventions 
 //common Resource Routes:
 // index - Show all listings
-// show - Show a single Lisitng
+// show - Show a single Listing
 // create- Show form to create new listing
 // store - Store new listing
 // edit - Show form to edit listing 
@@ -31,23 +32,40 @@ Route::get('/',[ListingController::class, 'index']);
 
 
 //Show create Form
-Route::get('/listings/create', [ListingController::class,'create']);
+Route::get('/listings/create', [ListingController::class,'create'])->middleware('auth');
 
 //Store Listings data
-Route::post('/listings', [ListingController::class,'store']);
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 
 //show edit Form
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
 //Edit submit to update
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //delete listing
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
+//show manage listing page
+Route::get('/listing/manage',[ListingController::class, 'manage'])->middleware('auth');
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class,'show']);
 
+
+//Show Register Create Form
+Route::get('/register',[UserController::class, 'create'])->middleware('guest');
+
+//create new user
+Route::post('/users',[UserController::class, 'store']);
+
+//log user out 
+Route::post('/logout',[UserController::class, 'logout'])->middleware('auth');
+
+//show login form
+Route::get('/login',[UserController::class, 'login'])->name('login')->middleware('guest');
+
+//login user
+Route::post('/users/authenticate',[UserController::class, 'authenticate']);
